backend/requirements.txt
fastapi
uvicorn
openai
pydantic
transformers
accelerate
PyMuPDF
python-dotenv
python-multipart